requirements.txt
selenium 
gspread==3.6.0 
oauth2client 
google-auth 
google-auth-oauthlib
google-auth-httplib2 
google-api-python-client 
pandas 
beautifulsoup4